requirements.txt
asgiref==3.8.1
certifi==2024.7.4
charset-normalizer==3.3.2
defusedxml==0.7.1
dj-rest-auth==6.0.0
Django==5.1
django-allauth==64.2.0
django-cors-headers==4.4.0
django-rest-auth==0.9.5
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
httpie==3.2.3
idna==3.8
inflection==0.5.1
markdown-it-py==3.0.0
mdurl==0.1.2
multidict==6.0.5
packaging==24.1
Pygments==2.18.0
PyJWT==2.9.0
PySocks==1.7.1
pytz==2024.1
PyYAML==6.0.2
requests==2.31.0
requests-toolbelt==1.0.0
rich==13.8.0
setuptools==74.0.0
six==1.16.0
sqlparse==0.5.1
uritemplate==4.1.1
urllib3==2.2.2
